[Change] Changed so that the results are saved in a CSV file when measurement is completed.
To reduce the number of times that the results are written to storage,
measurement results are no longer written to the CSV file during measurement,
but are instead stored in memory and written all at once when measurement is completed.

A pandas data frame is used to store the measurement results.

I made a similar change to the processing time measurement function (clc_elpsd_time),
but because it may affect other processes, I have commented it out for the time being and am using the existing process.